+++ FlowerWrapper/Models/Raw/fkapi_present_box.cs
@@ -13,20 +13,20 @@ namespace FlowerWrapper.Models.Raw
 	}
 	public class fkapi_userMailList
 	{
-		public int id { get; set; }
-		public int sendUserId { get; set; }
-		public int receiveUserId { get; set; }
+		public long id { get; set; }
+		public long sendUserId { get; set; }
+		public long receiveUserId { get; set; }
 		public string subject { get; set; }
 		public string body { get; set; }
-		public int sendUserTypeId { get; set; }
-		public int itemCategoryId { get; set; }
-		public int itemId { get; set; }
-		public int amount { get; set; }
-		public int characterLevelNum { get; set; }
-		public int characterLevelExperience { get; set; }
-		public int characterMaxEquipmentSlotCount { get; set; }
-		public int characterSkillLevelNum { get; set; }
-		public int status { get; set; }
+		public long sendUserTypeId { get; set; }
+		public long itemCategoryId { get; set; }
+		public long itemId { get; set; }
+		public long amount { get; set; }
+		public long characterLevelNum { get; set; }
+		public long characterLevelExperience { get; set; }
+		public long characterMaxEquipmentSlotCount { get; set; }
+		public long characterSkillLevelNum { get; set; }
+		public long status { get; set; }
 		public string receivedDate { get; set; }
 		public string limitDate { get; set; }
 	}
